test(card): add accessibility tests
