+++ cypress/support/accessibility/a11y-utils.js
@@ -71,6 +71,7 @@ export default (from, end) => {
       !prepareUrl[0].startsWith("dialog-full-screen") &&
       !prepareUrl[0].startsWith("verticalmenu") &&
       !prepareUrl[0].startsWith("message") &&
+      !prepareUrl[0].startsWith("card") &&
       !prepareUrl[0].endsWith("test")
     ) {
       urlList.push([prepareUrl[0], prepareUrl[1]]);
